Move existing example.dart to multi_isolate.dart and create a simpler 'Hello World' server
